requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.11.1
async-timeout==5.0.1
asyncpg==0.29.0
bcrypt==5.0.0
boto3==1.40.74
botocore==1.40.74
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
coverage==7.12.0
cryptography==46.0.3
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.104.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jmespath==1.0.1
limits==5.6.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
passlib==1.7.4
pillow==12.0.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
Pygments==2.19.2
pytest==9.0.1
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
s3transfer==0.14.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
stripe==7.7.0
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wrapt==2.0.1
